fixed casting error

+++ src/GpioMotor.h
@@ -4,7 +4,6 @@
 
 #ifndef GPIOCPP_GPIOWHEEL_H
 #define GPIOCPP_GPIOWHEEL_H
-#include <inttypes.h>
 namespace pi_ln298n{
 
 using GpioId = unsigned int;
@@ -20,7 +19,7 @@ protected:
 
     void forward(float effortPerc);
     void backward(float effortPerc);
-    uint8_t effort2pwm(float effortPerc);
+    int effort2pwm(float effortPerc);
     static unsigned int _nInstances;
     GpioId _forward,_backward,_enable;
 };
